requirements.txt
# Requirements for Real Estate Chatbot

# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.6.0
pandas>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0

# Data processing
numpy>=1.24.0

# UI enhancements
streamlit-chat>=0.1.1

# Development tools
black>=23.0.0
pylint>=2.17.0